Revert "adapted to moved SeverityAndIssue classes"
This reverts commit a62bbbe84d5676492a0b4b82b65ffb2becf2952f.
